add authentication

+++ routers/booking.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import Session
 from crud_func.bookingfun import book_new_flight, get_booking, delete_booking_by_id, get_all_booking
 from typing import List
 from routers.login import oauth2_scheme
+from crud_func.login import get_authorize
 
 router = APIRouter(tags = ['booking'])
 
@@ -23,11 +24,13 @@ def get_booking_by_id(id:int , db: Session = Depends(get_db)):
 
 @router.get("/booking", response_model=List[ShowBookedFlight], status_code=status.HTTP_200_OK)
 def get__all_booking(db: Session = Depends(get_db), token:str=Depends(oauth2_scheme)):
-    ref = get_all_booking(db=db)
-    return ref
+    if get_authorize(token,db):
+        ref = get_all_booking(db=db)
+        return ref
 
 
-@router.delete("/booking/{id}" , response_model=ShowBookedFlight, status_code=status.HTTP_200_OK)
+@router.delete("/booking/{id}", status_code=status.HTTP_200_OK)
 def delete_booking(id:int , db:Session= Depends(get_db), token:str=Depends(oauth2_scheme)):
-    obj = delete_booking_by_id(id=id, db=db)
-    return obj
+    if get_authorize(token, db):
+        obj = delete_booking_by_id(id=id, db=db)
+        return "booking deleted successfully"
