+++ routers/flightroute.py
@@ -5,14 +5,16 @@ from sqlalchemy.orm import Session
 from crud_func.flightroutefun import get_flight_route, new_flight_route, delete_flight_route, get_all_flight_route
 from typing import List
 from routers.login import oauth2_scheme
+from crud_func.login import get_authorize
 
 router = APIRouter( tags= ['flightroute'])
 
 
 @router.post("/flightroute", status_code=status.HTTP_201_CREATED)
 def create_flightroute(flight: CreateFlightRoute, db: Session = Depends(get_db), token:str=Depends(oauth2_scheme)):
-    new_flight = new_flight_route(flight=flight, db=db)
-    return "flight route created"
+    if get_authorize(token, db):
+        new_flight = new_flight_route(flight=flight, db=db)
+        return "flight route created"
 
 
 
@@ -24,8 +26,9 @@ def get_flightroute(db: Session = Depends(get_db)):
 
 @router.delete("/flightroute/{id}",status_code=status.HTTP_200_OK)
 def delete_flightroute(id:int , db:Session= Depends(get_db), token:str=Depends(oauth2_scheme)):
-    obj = delete_flight_route(id=id, db=db)
-    return "Successfully deleted"
+    if get_authorize(token, db):
+        obj = delete_flight_route(id=id, db=db)
+        return "Successfully deleted"
